requirements.txt
Flask
pymysql
python-dotenv
pandas
flask-cors